Moved profile to user folder
