+++ src/App.jsx
@@ -9,7 +9,7 @@ import { useCookies } from 'react-cookie'
 import Login from './components/pages/login/Login'
 import Cart from './components/pages/cart/Cart'
 import ProductPage from './components/pages/productPage/ProductPage'
-import Profile from './components/pages/profile/Profile'
+import Profile from './components/pages/user/profile/Profile'
 import LandingPage from './components/pages/onBoarding/LandingPage'
 import SigninOptions from './components/pages/onBoarding/SigninOptions'
 import Favourites from './components/pages/favourites/Favourites'
